ontology-management-service/pyproject.toml
[tool.poetry]
name = "oms-monolith"
version = "3.0.0"
description = "Ontology Management System - Ultimate Monolith"
authors = ["OMS Team"]
packages = [{ include = "*" }]

[tool.poetry.dependencies]
python = ">=3.11"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
pydantic = "2.5.0"
pydantic-settings = "2.1.0"
asyncpg = "0.29.0"
motor = "3.3.2"
aiosqlite = "0.19.0"
sqlalchemy = "2.0.23"
alembic = "1.12.1"
psycopg2-binary = "2.9.9"
nats-py = "2.6.0"
celery = "5.3.4"
kombu = "5.3.2"
redis = "5.0.1"
hiredis = "2.3.2"
httpx = {extras = ["http2"], version = ">=0.25.2,<0.26"}
aiohttp = "3.9.1"
python-jose = {extras = ["cryptography"], version = "3.3.0"}
python-multipart = "0.0.6"
cryptography = ">=41.0.7"
pyjwt = "2.8.0"
passlib = {extras = ["bcrypt"], version = "1.7.4"}
strawberry-graphql = {extras = ["fastapi"], version = "0.209.0"}
graphene = "3.3"
boto3 = "1.29.7"
botocore = "1.32.7"
prometheus-client = "0.19.0"
opentelemetry-api = "1.24.0"
opentelemetry-sdk = "1.24.0"
opentelemetry-instrumentation-fastapi = "0.45b0"
opentelemetry-instrumentation-grpc = "0.45b0"
prometheus-fastapi-instrumentator = "6.1.0"
grpcio = "1.60.0"
grpcio-tools = "1.60.0"
terminusdb-client = "10.2.6"
python-dotenv = "1.0.0"
pendulum = ">=3.0.0"
python-dateutil = "2.8.2"
click = "8.1.7"
punq = "0.7.0"
dependency-injector = "4.41.0"
sentry-sdk = {extras = ["fastapi"], version = "1.38.0"}
structlog = "23.2.0"
arq = "0.25.0"
minio = "7.1.17"
jsonschema = "4.20.0"
cerberus = "1.3.5"
typer = "0.9.0"
email-validator = ">=2.1.1"
python-magic = "0.4.27"
chardet = "5.2.0"
cloudevents = "1.10.1"
pybreaker = "1.0.2"
backoff = "2.2.1"
psutil = "5.9.6"
aiodataloader = "0.2.1"
networkx = ">=3.0"
numpy = ">=1.24.0"
scikit-learn = ">=1.3.0"
tiktoken = ">=0.5.0"
cohere = ">=4.0.0"
transformers = ">=4.30.0"
sentence-transformers = ">=2.2.0"
google-cloud-aiplatform = ">=1.35.0"
anthropic = ">=0.7.0"
opentelemetry-exporter-jaeger = "1.21.0"
opentelemetry-instrumentation-redis = "0.45b0"
opentelemetry-instrumentation-asyncio = "0.45b0"
opentelemetry-instrumentation-requests = "0.45b0"
jsonpatch = "1.33"
jsondiff = "2.0.0"
python-json-logger = "2.0.7"
zstandard = ">=0.21.0"

[tool.poetry.group.dev.dependencies]
import-linter = "^1.12.0"
black = "^23.0.0"
isort = "^5.12.0"
mypy = "^1.7.0"
ruff = "^0.1.0"

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=61.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_optional = true
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_unused_imports = true
show_error_codes = true

# Per-module settings
[[tool.mypy.overrides]]
module = [
    "tests.*",
    "migrations.*",
    "_archive_tests.*"
]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "httpx.*",
    "fastapi.*",
    "uvicorn.*",
    "pydantic.*",
    "strawberry.*",
    "nats.*",
    "asyncpg.*",
    "pytest.*"
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py39"

select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "Q",   # flake8-quotes
    "SIM", # flake8-simplify
    "ICN", # flake8-import-conventions
    "PIE", # flake8-pie
    "T20", # flake8-print
    "PYI", # flake8-pyi
    "PT",  # flake8-pytest-style
    "RET", # flake8-return
    "SLF", # flake8-self
    "ARG", # flake8-unused-arguments
    "PTH", # flake8-use-pathlib
    "ERA", # eradicate
    "PGH", # pygrep-hooks
    "PL",  # pylint
    "TRY", # tryceratops
    "FURB", # refurb
    "LOG",  # flake8-logging
    "RUF",  # ruff-specific
]

ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "PLR0913", # too many arguments
    "PLR0912", # too many branches
    "PLR0915", # too many statements
    "PLR2004", # magic value used in comparison
    "RET504",  # unnecessary variable assignment
    "T201",    # print statements (allowed in scripts)
    "T203",    # pprint statements
]

exclude = [
    ".git",
    "__pycache__",
    "*.pyc",
    ".venv",
    "_archive_tests",
    "migrations",
]

[tool.ruff.per-file-ignores]
"tests/**/*.py" = [
    "S101",   # use of assert
    "ARG",    # unused function args
    "PLR2004", # magic value
]
"scripts/**/*.py" = [
    "T201",   # print statements allowed in scripts
]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
# asyncio_mode = "strict"  # Removed - not compatible with current pytest-asyncio
markers = [
    "asyncio: marks tests as asyncio tests",
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "e2e: marks tests as end-to-end tests",
    "real: marks tests that use real services (no mocks)",
    "etag: marks tests as ETag caching tests (deselect with '-m \"not etag\"')",
    "unit: marks tests as unit tests",
    "benchmark: marks tests as performance benchmarks",
    # New markers for enhanced testing
    "security: marks tests as security-related tests",
    "override: marks tests related to override approval system",
    "config: marks tests related to configuration system",
    "audit: marks tests related to audit and logging",
    "validation: marks tests related to validation and bypass",
    "authorization: marks tests related to authorization and permissions",
    "middleware: marks tests related to middleware functionality",
    "performance: marks tests related to performance characteristics",
    "compliance: marks tests related to compliance and reporting",
]
